Use LF to terminate lines instead of CRLF.
The documentation was changed to state that LF is now used:
https://git.tartarus.org/?p=simon/putty.git;a=commitdiff;h=a56b79b598902c7393fd79c63373a8b4eba4833c

The commit message states:

> But I noticed that we're actually outputting \n by preference, whereas
> the spec said we prefer \r\n. I'd rather change the docs than the
> code.

Despite this, the current PuTTYgen 0.74 release outputs CRLF terminated
lines. The current development snapshot outputs LF terminated lines.

The change appears to be that the file is now opened in binary mode (and
an intermediate buffer is now used):
https://git.tartarus.org/?p=simon/putty.git;a=commitdiff;h=9cc586e605e3db1d6c2294f209da26e633046f3f